[REF] InternalImageStorage 클래스 별도 파일로 분리
