+++ core/localstorage/src/main/java/com/hilingual/core/localstorage/DiaryTempManagerImpl.kt
@@ -1,23 +1,15 @@
 package com.hilingual.core.localstorage
 
-import android.content.Context
 import android.net.Uri
-import androidx.core.net.toUri
 import androidx.datastore.core.DataStore
-import androidx.datastore.core.IOException
 import androidx.datastore.preferences.core.MutablePreferences
 import androidx.datastore.preferences.core.Preferences
 import androidx.datastore.preferences.core.booleanPreferencesKey
 import androidx.datastore.preferences.core.edit
 import androidx.datastore.preferences.core.stringPreferencesKey
-import dagger.hilt.android.qualifiers.ApplicationContext
-import kotlinx.coroutines.Dispatchers
+import com.hilingual.core.localstorage.util.InternalImageStorage
 import kotlinx.coroutines.flow.first
-import kotlinx.coroutines.withContext
-import java.io.File
-import java.io.FileOutputStream
 import java.time.LocalDate
-import java.util.UUID
 import javax.inject.Inject
 
 class DiaryTempManagerImpl @Inject constructor(
@@ -92,38 +84,6 @@ class DiaryTempManagerImpl @Inject constructor(
             if (oldImageUriString != null && oldImageUriString != savedUri.toString()) {
                 imageStorage.deleteImageFromInternal(oldImageUriString)
             }
-        }
-    }
-}
-
-class InternalImageStorage @Inject constructor(@ApplicationContext private val context: Context) {
-    suspend fun saveImageToInternal(imageUri: Uri): Uri? = withContext(Dispatchers.IO) {
-        val directory = File(context.filesDir, "diary_images").apply { mkdirs() }
-        val destFile = File(directory, "img_${UUID.randomUUID()}.jpg")
-
-        try {
-            context.contentResolver.openInputStream(imageUri).use { input ->
-                FileOutputStream(destFile).use { output ->
-                    input?.copyTo(output)
-                }
-            } ?: throw IOException()
-
-            Uri.fromFile(destFile)
-        } catch (e: Exception) {
-            e.printStackTrace()
-            if (destFile.exists()) destFile.delete()
-            null
-        }
-    }
-
-    suspend fun deleteImageFromInternal(imageUriString: String?) = withContext(Dispatchers.IO) {
-        if (imageUriString == null) return@withContext
-
-        val path = imageUriString.toUri().path ?: return@withContext
-
-        val file = File(path)
-        if (file.exists()) {
-            file.delete()
         }
     }
 }
