+++ core/localstorage/src/main/java/com/hilingual/core/localstorage/di/LocalStorageModule.kt
@@ -21,7 +21,6 @@ import androidx.datastore.dataStore
 import androidx.datastore.preferences.preferencesDataStore
 import com.hilingual.core.localstorage.DiaryTempManager
 import com.hilingual.core.localstorage.DiaryTempManagerImpl
-import com.hilingual.core.localstorage.InternalImageStorage
 import com.hilingual.core.localstorage.TokenManager
 import com.hilingual.core.localstorage.TokenManagerImpl
 import com.hilingual.core.localstorage.UserInfoManager
@@ -29,6 +28,7 @@ import com.hilingual.core.localstorage.UserInfoManagerImpl
 import com.hilingual.core.localstorage.constant.DataStoreConstant
 import com.hilingual.core.localstorage.model.UserPreferences
 import com.hilingual.core.localstorage.serializer.UserPreferencesSerializer
+import com.hilingual.core.localstorage.util.InternalImageStorage
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
